Add margins back into StringWidget
